requirements.txt
joblib==1.1.0
numpy==1.21.4
opencv-python==4.5.4.58
pandas==1.1.5
pickle-mixin==1.0.2
pickleshare==0.7.5
Pillow==8.4.0
python-dateutil==2.8.2
pytz==2021.3
scikit-learn==0.22.2
scipy==1.7.2
six==1.16.0
threadpoolctl==3.0.0
